ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
pymysql==1.1.0
sqlalchemy==2.0.23
minio==7.2.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
